Add files via upload
Adding a graphical flowchart for human contributors, revised with respect to the problems/inconsistencies described in #9 (https://github.com/NFDI4Chem/rxno/issues/9), to put newly added reactions in the correct parent class. A file highlighting the changes made is also provided for comparison.

Addition of a new graphical flowchart for skeleton determination in addition to the existing written version.
Still open questions regarding the application of this new skeleton flowchart:
Which substance is to be defined as starting material, in case of several "equal" reaction partners?
The rules of the scheme must always be applied once to both the starting material and the reaction product to determine the skeleton or only to the starting material?